refactor: moved InputForPhone back to stories
