+++ src/stories/1-Input.stories.tsx
@@ -4,7 +4,7 @@
  * File Created: Wednesday, 8th July 2020 1:55:18 am
  * Author: Gabriel Ulloa (gabriel@inventures.cl)
  * -----
- * Last Modified: Friday, 21st August 2020 10:40:35 am
+ * Last Modified: Monday, 24th August 2020 3:21:50 pm
  * Modified By: Esperanza Horn (esperanza@inventures.cl)
  * -----
  * Copyright 2019 - 2020 Incrementa Ventures SpA. ALL RIGHTS RESERVED
@@ -12,10 +12,9 @@
  * -----
  * Inventures - www.inventures.cl
  */
-import React, { useCallback } from 'react';
+import React, { useCallback, useState, ChangeEvent } from 'react';
 import { number, text } from '@storybook/addon-knobs';
 import { Input } from '../components/input';
-
 import { InputStatus, useInput } from '../hooks/useInput.hooks';
 import { rutFormat } from 'rut-helpers';
 import {
@@ -24,12 +23,26 @@ import {
   RutFormatValidator,
   EmailValidator,
   RutValidator,
+  LengthValidator,
+  NumericValidator,
 } from '../hooks/validators';
 import { LatinEmailFormatter } from '../hooks/formatters';
-import { CountryType, InputForPhoneComponent } from '../components';
+import Select from '@material-ui/core/Select';
+import FormControl from '@material-ui/core/FormControl';
+import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
+import MenuItem from '@material-ui/core/MenuItem';
 
 export default {
   title: 'Input',
+};
+export type CountryType = {
+  id: number;
+  countryName: string;
+  countryDigitLength: number;
+  countryPrefix: number;
+};
+type InputforPhoneProps = {
+  possibleCountries: CountryType[];
 };
 
 export const Base = () => <Input />;
@@ -126,29 +139,109 @@ export const InputForEmail = () => {
   );
 };
 
-export const InputForPhone = () => {
-  const possibleCountries = [
-    {
-      id: 1,
-      countryName: 'Chile',
-      countryDigitLength: 9,
-      countryPrefix: 56,
-    } as CountryType,
-    {
-      id: 2,
-      countryName: 'Peru',
-      countryDigitLength: 9,
-      countryPrefix: 51,
-    } as CountryType,
-    {
-      id: 3,
-      countryName: 'USA',
-      countryDigitLength: 11,
-      countryPrefix: 1,
-    } as CountryType,
-  ];
+const possibleCountriesDummy = [
+  {
+    id: 1,
+    countryName: 'Chile',
+    countryDigitLength: 9,
+    countryPrefix: 56,
+  } as CountryType,
+  {
+    id: 2,
+    countryName: 'Peru',
+    countryDigitLength: 9,
+    countryPrefix: 51,
+  } as CountryType,
+  {
+    id: 3,
+    countryName: 'USA',
+    countryDigitLength: 11,
+    countryPrefix: 1,
+  } as CountryType,
+];
 
-  return <InputForPhoneComponent possibleCountries={possibleCountries} />;
+export const InputForPhone = (props: InputforPhoneProps) => {
+  const classes = useStyles();
+
+  //const { possibleCountries } = props;
+  const possibleCountries = possibleCountriesDummy;
+  const [country, setCountry] = useState(possibleCountries[0]);
+
+  const nonNumeric = text(
+    'Teléfono no numérico error',
+    '¡Ups! Recuerda incluir sólo números',
+  );
+  const required = text('Teléfono requerido error', 'Teléfono Requerido');
+  const incompleteNumber = text(
+    'Teléfono incompleto error',
+    '¡Ojo! Debes incluir la cantidad correcta de digitos para tu país',
+  );
+  const debounceTime = number('Debounce time (ms)', 800);
+
+  const [
+    value,
+    setValue,
+    status,
+    errors,
+    handleBlur,
+    updateMaxLength,
+  ] = useInput('', {
+    validators: [
+      required && new RequiredValidator(required),
+      nonNumeric && new NumericValidator(nonNumeric),
+      incompleteNumber && new LengthValidator(incompleteNumber),
+    ].filter(Boolean) as Validator<string>[],
+    debounceTime,
+    maxLength: country.countryDigitLength,
+  });
+
+  // when change of country, update the country and run callback functionon useInput to update max length
+  const handleChange = useCallback(
+    (event: ChangeEvent<{ value: unknown }>) => {
+      const newCountry = event.target.value as CountryType;
+      setCountry(newCountry);
+      updateMaxLength(newCountry.countryDigitLength);
+    },
+    [updateMaxLength, setCountry],
+  );
+
+  const handleWrite = useCallback(
+    (e) => {
+      setValue(String(e.target.value));
+    },
+    [setValue],
+  );
+
+  return (
+    <FormControl className={classes.formControl}>
+      <Select value={country} onChange={handleChange}>
+        {possibleCountries.map((item, index) => (
+          // TODO: understand why it complains about using an object - yet compiles and works fine
+          <MenuItem key={index} value={item}>
+            {item.countryName} (+{item.countryPrefix})
+          </MenuItem>
+        ))}
+      </Select>
+      <Input
+        value={value}
+        onChange={handleWrite}
+        onBlur={handleBlur}
+        error={status === InputStatus.ERROR}
+        helperText={errors[0]}
+        label={`Phone* (${debounceTime}ms)`}
+      />
+    </FormControl>
+  );
 };
 
 Base.storyName = 'Base element';
+
+const useStyles = makeStyles((theme: Theme) =>
+  createStyles({
+    formControl: {
+      margin: theme.spacing(1),
+      minWidth: 100,
+      flexDirection: 'row',
+    },
+  }),
+);
